BUILD : 1.7.10 (843) BUGS : NOTES : Fixed various compile warnings on AMD64 systems
git-svn-id: svn://svn.anope.org/anope/trunk@843 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@596 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/config.c
@@ -687,7 +687,7 @@ int parse_directive(Directive * d, char *dir, int ac, char *av[MAXPARAMS],
 {
     int retval = 1;
     int i;
-    int val;
+    long val;
     int optind;
 
     if (stricmp(dir, d->name) != 0)
